bet-maker/requirements.txt
aiohttp==3.8.6
SQLAlchemy==2.0.22
asyncpg==0.28.0
sqlalchemy-utils==0.41.1
APScheduler==3.10.4
fastapi==0.104.1
uvicorn==0.23.2
pydantic-settings==2.0.3
alembic==1.12.1